fix: disable cut command in table cell

+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/copy_and_paste/custom_cut_command.dart
@@ -1,8 +1,8 @@
-import 'package:flutter/material.dart';
-
 import 'package:appflowy/plugins/document/presentation/editor_plugins/plugins.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/table/table_operations/table_operations.dart';
 import 'package:appflowy/shared/clipboard_state.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
+import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 /// cut.
@@ -42,6 +42,11 @@ CommandShortcutEventHandler _cutCommandHandler = (editorState) {
     if (node == null) {
       return KeyEventResult.handled;
     }
+    // prevent to cut the node that is selecting the table.
+    if (node.parentTableNode != null) {
+      return KeyEventResult.skipRemainingHandlers;
+    }
+
     final transaction = editorState.transaction;
     transaction.deleteNode(node);
     final nextNode = node.next;
